refactor: centralize corner items
one way to centralize corner items is to add dummy items as new corner items,
and after you scroll to it you can forcefully scroll back to a valid item.

now passing the prop alwaysCentralized all selected items will be centralized,
leveraging dummy corner items, with these changes
when user scrolls to one of the corners he will be scrolled back
to the item that is directly before the dummy item.